Refactored new gpx schema and changed route/blog_post into a many to many

+++ src/main/resources/db/migration/V10__add_coordinate_table.sql
@@ -3,8 +3,8 @@ CREATE TABLE if NOT EXISTS coordinate (
     lat FLOAT NOT NULL,
     lng FLOAT NOT NULL,
     altitude FLOAT,
-    route_id int,
+    route_id INT,
     created_date TIMESTAMP NOT NULL,
-    heart_rate bigint,
+    heart_rate INT,
     CONSTRAINT fk_route FOREIGN KEY(route_id) REFERENCES route(id)
 );
